+++ packages/dialtone-vue3/components/toggle/toggle_default.story.vue
@@ -5,6 +5,7 @@
     :size="$attrs.size"
     :show-icon="$attrs.showIcon"
     :label-class="$attrs.labelClass"
+    :wrapper-class="$attrs.wrapperClass"
     :label-child-props="$attrs.labelChildProps"
     :toggle-on-click="$attrs.toggleOnClick"
     @change="$attrs.onChange"