+++ packages/dialtone-vue3/directives/scrollbar_directive/scrollbar.js
@@ -4,9 +4,10 @@ export const DtScrollbarDirective = {
   name: 'dt-scrollbar-directive',
   install (app) {
     OverlayScrollbars.plugin(ClickScrollPlugin);
+    const instances = new WeakMap();
     app.directive('dt-scrollbar', {
       mounted (el, binding) {
-        OverlayScrollbars({
+        const os = OverlayScrollbars({
           target: el,
           elements: {
             viewport: el.children[0],
@@ -20,6 +21,10 @@ export const DtScrollbarDirective = {
         });
         el.setAttribute('data-overlayscrollbars-initialize', true);
         el.classList.add('d-scrollbar');
+        instances.set(el, os);
+      },
+      unmounted (el) {
+        instances.get(el).destroy();
       },
     });
   },